rename event and user onboarding files
