+++ src/components/views/Dashboard.js
@@ -1,5 +1,5 @@
 import React from "react";
-import Onboarding from "../templates/Onboarding";
+import EventOnboarding from "../templates/Onboarding";
 import { useDispatch } from "react-redux";
 import { fetchAllEvents } from "../../store/events/actions";
 
@@ -10,7 +10,7 @@ const Dashboard = () => {
     dispatch(fetchAllEvents());
   }, []);
 
-  return <Onboarding user="Mildred Pascal" />;
+  return <EventOnboarding user="Mildred Pascal" />;
 };
 
 export default Dashboard;
