+++ src/components/views/SignupPage.jsx
@@ -1,10 +1,10 @@
 import React from "react";
 import image from "../../assets/Signup.png";
-import { Unboarding } from "../templates";
+import { UserOnboarding } from "../templates";
 
 const SignupPage = () => {
   return (
-    <Unboarding
+    <UserOnboarding
       ctaText="Sign Up"
       imageType={image}
       imageText="Sign Up now!!"
